+++ CNR_FURNITURE/src/main/resources/static/js/work/work.js
@@ -1,0 +1,56 @@
+//페이지가 로드될 때
+$(document).ready(function() {
+	// 해당 표가 보이지 않도록 한다.
+  $('#selectWorker').css('display', 'none');
+  
+	 // 자재투입내역 버튼을 누를 경우
+	$('#selectMaterialInputBtn').click(function(){
+	  $('#selectMaterialInput').css('display', 'block');
+	  $('#selectWorker').css('display', 'none');
+	  
+	  $('#selectMaterialInputBtn').css('background', '#374056');
+	  $('#selectMaterialInputBtn').css('color', '#fff');
+	  
+	  $('#selectWorkerBtn').css('color', 'inherit');
+	  $('#selectWorkerBtn').css('background', 'inherit');
+	});
+	 
+	// 작업자 버튼을 누를 경우
+	$('#selectWorkerBtn').click(function(){
+	  $('#selectMaterialInput').css('display', 'none');
+	  $('#selectWorker').css('display', 'block');
+	  
+	  $('#selectWorkerBtn').css('background', '#374056');
+	  $('#selectWorkerBtn').css('color', '#fff');
+	  
+	  $('#selectMaterialInputBtn').css('color', 'inherit');
+	  $('#selectMaterialInputBtn').css('background', 'inherit');
+	});
+	
+	// reset 이미지가 담긴 박스를 클릭할 경우 검색바의 input 값 초기화
+	$('.reset').click(function() {
+  	$('#find_work_company').val("");
+  	$('#find_work_item').val("");
+  	$('#workStartDate').val("");
+  	$('#workEndDate').val("");
+  	$('#find_work_instruction').val("");
+  	$('#find_work_processInfo').val("");
+  	$('#find_work_id').val("");
+	});
+	
+	
+}); // 페이지 처리
+	
+	
+	
+
+	
+	
+	
+	
+
+
+
+
+
+
